+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    int health { get; set; }
-    int maxHealth { get; set; }
+    float health { get; set; }
+    float maxHealth { get; set; }
     void DamageEntity(int dmg);
     void DestroyEntity();
     void HealEntity(int amounnt);
